Rename AsyncInMemoryEventStorageEngine to InMemoryEventStorageEngine
Rename AsyncInMemoryEventStorageEngine to InMemoryEventStorageEngine.
This clarifies that this component replace the old
InMemoryEventStorageEngine entirely.
