+++ eventsourcing/src/test/java/org/axonframework/eventsourcing/configuration/EventSourcingConfigurationDefaultsTest.java
@@ -28,7 +28,7 @@ import org.axonframework.eventsourcing.eventstore.EventStore;
 import org.axonframework.eventsourcing.eventstore.SimpleEventStore;
 import org.axonframework.eventsourcing.eventstore.Tag;
 import org.axonframework.eventsourcing.eventstore.TagResolver;
-import org.axonframework.eventsourcing.eventstore.inmemory.AsyncInMemoryEventStorageEngine;
+import org.axonframework.eventsourcing.eventstore.inmemory.InMemoryEventStorageEngine;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.*;
 
@@ -62,7 +62,7 @@ class EventSourcingConfigurationDefaultsTest {
         Configuration resultConfig = configurer.build();
 
         assertInstanceOf(AnnotationBasedTagResolver.class, resultConfig.getComponent(TagResolver.class));
-        assertInstanceOf(AsyncInMemoryEventStorageEngine.class,
+        assertInstanceOf(InMemoryEventStorageEngine.class,
                          resultConfig.getComponent(EventStorageEngine.class));
         EventStore eventStore = resultConfig.getComponent(EventStore.class);
         assertInstanceOf(SimpleEventStore.class, eventStore);
