+++ eventsourcing/src/test/java/org/axonframework/eventsourcing/eventstore/inmemory/InMemoryEventStorageEngineTest.java
@@ -16,93 +16,18 @@
 
 package org.axonframework.eventsourcing.eventstore.inmemory;
 
-import org.axonframework.eventhandling.EventMessage;
-import org.axonframework.eventhandling.EventTestUtils;
-import org.axonframework.eventhandling.TrackedEventMessage;
-import org.axonframework.eventsourcing.eventstore.EventStorageEngineTest;
-import org.axonframework.messaging.unitofwork.DefaultUnitOfWork;
-import org.axonframework.messaging.unitofwork.UnitOfWork;
-import org.junit.jupiter.api.*;
-
-import java.util.Optional;
-import java.util.OptionalLong;
-import java.util.stream.Stream;
-
-import static org.junit.jupiter.api.Assertions.*;
+import org.axonframework.eventsourcing.eventstore.SimpleEventStore;
+import org.axonframework.eventsourcing.eventstore.StorageEngineTestSuite;
 
 /**
- * Test class validating the {@link LegacyInMemoryEventStorageEngine}.
+ * Test class validating the {@link SimpleEventStore} together with the {@link InMemoryEventStorageEngine}.
  *
- * @author Rene de Waele
+ * @author Steven van Beelen
  */
-class InMemoryEventStorageEngineTest extends EventStorageEngineTest {
+class InMemoryEventStorageEngineTest extends StorageEngineTestSuite<InMemoryEventStorageEngine> {
 
-    private static final EventMessage<Object> TEST_EVENT = EventTestUtils.asEventMessage("test");
-
-    private LegacyInMemoryEventStorageEngine testSubject;
-
-    @BeforeEach
-    void setUp() {
-        testSubject = new LegacyInMemoryEventStorageEngine();
-        setTestSubject(testSubject);
-    }
-
-    @Test
-    void publishedEventsEmittedToExistingStreams() {
-        Stream<? extends TrackedEventMessage<?>> stream = testSubject.readEvents(null, true);
-        testSubject.appendEvents(TEST_EVENT);
-
-        assertTrue(stream.findFirst().isPresent());
-    }
-
-    @Test
-    void publishedEventsEmittedToExistingStreams_WithOffset() {
-        testSubject = new LegacyInMemoryEventStorageEngine(1);
-        Stream<? extends TrackedEventMessage<?>> stream = testSubject.readEvents(null, true);
-        testSubject.appendEvents(TEST_EVENT);
-
-        Optional<? extends TrackedEventMessage<?>> optionalResult = stream.findFirst();
-        assertTrue(optionalResult.isPresent());
-        OptionalLong optionalResultPosition = optionalResult.get().trackingToken().position();
-        assertTrue(optionalResultPosition.isPresent());
-        assertEquals(1, optionalResultPosition.getAsLong());
-    }
-
-    @Test
-    void eventsAreStoredOnCommitIfCurrentUnitOfWorkIsActive() {
-        UnitOfWork<EventMessage<Object>> unitOfWork = DefaultUnitOfWork.startAndGet(TEST_EVENT);
-
-        // when _only_ publishing...
-        testSubject.appendEvents(TEST_EVENT);
-
-        // then there are no events in the storage engine, since the UnitOfWork is not committed yet.
-        Stream<? extends TrackedEventMessage<?>> eventStream = testSubject.readEvents(null, true);
-        assertEquals(0L, eventStream.count());
-
-        // When rolling back the UnitOfWork...
-        unitOfWork.commit();
-
-        // then there are *still* no events in the storage engine.
-        eventStream = testSubject.readEvents(null, true);
-        assertEquals(1L, eventStream.count());
-    }
-
-    @Test
-    void eventsAreNotStoredWhenTheUnitOfWorkIsRolledBackIfCurrentUnitOfWorkIsActive() {
-        UnitOfWork<EventMessage<Object>> unitOfWork = DefaultUnitOfWork.startAndGet(TEST_EVENT);
-
-        // when _only_ publishing...
-        testSubject.appendEvents(TEST_EVENT);
-
-        // then there are no events in the storage engine, since the UnitOfWork is not committed yet.
-        Stream<? extends TrackedEventMessage<?>> eventStream = testSubject.readEvents(null, true);
-        assertEquals(0L, eventStream.count());
-
-        // When rolling back the UnitOfWork...
-        unitOfWork.rollback();
-
-        // then there are *still* no events in the storage engine.
-        eventStream = testSubject.readEvents(null, true);
-        assertEquals(0L, eventStream.count());
+    @Override
+    protected InMemoryEventStorageEngine buildStorageEngine() {
+        return new InMemoryEventStorageEngine();
     }
 }
